modularize_Code
modularize the code
