+++ .gitignore
@@ -1,1 +1,5 @@
 p3_Summer_Camp_env/
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
